image-to-knowledge/requirements.txt
openai>=1.0.0
requests>=2.28.0
python-dotenv>=1.0.0
Pillow>=10.0.0
PyMuPDF>=1.24.0
